refactor: append Utils to all main utility files
