+++ frontend/__tests__/components/Explore/CommunityBox.test.tsx
@@ -1,5 +1,4 @@
 import CommunityBox from "@/components/Explore/CommunityBox";
-import userEvent from "@testing-library/user-event";
 import { render, screen } from "@testing-library/react";
 import "@testing-library/jest-dom";
 import { CommunityExploreAPI } from "@/types/api";
