+++ frontend/__tests__/components/UserProfileProvider.test.tsx
@@ -1,7 +1,6 @@
 import UserProfileProvider from "@/components/UserProfileProvider";
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
-// import userEvent from "@testing-library/user-event";
 import { UserProfile } from "@/types/user";
 import React, { act } from "react";
 import { UserContext } from "@/contexts/user";
