+++ frontend/src/components/EditProfile/EditProfileFormWrapper.tsx
@@ -8,7 +8,7 @@ import {
 } from "@/types/form";
 import { User, UserProfile } from "@/types/user";
 import EditProfileForm from "@/components/EditProfile/EditProfileForm";
-import { createOnSubmit, getFormData } from "@/utils/form";
+import { createOnSubmit, getFormData } from "@/utils/formUtils";
 
 const EditProfileFormWrapper = () => {
   const { userProfile } = useContext(UserProfileContext);
