+++ frontend/src/utils/stringUtils.ts
